MailAgent/requirements.txt
openai
python-dotenv>=0.19.0
requests 